Add separate resourceful endpoints for transactions
If we had an existing client in production we'd probably want to
make this change more incrementally- add the new endpoint, leaving the
old one intact, make the client use the new endpoint, then pull
the functionality from the generic actions endpoint. But since this
is a toy example, let's just pull it.

I'd probably go a setp further and pull all the instantiation of
different kinds of AccountActionRequests into a factory